merge 3/24/2021

+++ code/modules/research/xenobiology/xenobiology.dm
@@ -943,6 +943,12 @@
 
 	qdel(src)
 
+/obj/item/slimepotion/slime/slimeradio/ncr
+	name = "bluespace radio potion"
+	desc = "A strange chemical that grants those who ingest it the ability to broadcast and receive subscape radio waves."
+	icon = 'icons/obj/chemical.dmi'
+	icon_state = "potgrey"
+
 /obj/item/slimepotion/slime/slimeradio
 	name = "bluespace radio potion"
 	desc = "A strange chemical that grants those who ingest it the ability to broadcast and receive subscape radio waves."
@@ -963,6 +969,23 @@
 	to_chat(user, "<span class='notice'>You feed the potion to [M].</span>")
 	to_chat(M, "<span class='notice'>Your mind tingles as you are fed the potion. You can hear radio waves now!</span>")
 	var/obj/item/implant/radio/slime/imp = new
+	imp.implant(M, user)
+	qdel(src)
+
+/obj/item/slimepotion/slime/slimeradio/ncr/attack(mob/living/M, mob/user)
+	if(!isanimal(M))
+		to_chat(user, "<span class='warning'>[M] is too complex for the potion!</span>")
+		return
+	if(!M.can_be_implanted())
+		to_chat(user, "<span class='warning'>[M] is incompatible with the potion!</span>")
+		return
+	if(M.stat)
+		to_chat(user, "<span class='warning'>[M] is dead!</span>")
+		return
+
+	to_chat(user, "<span class='notice'>You feed the potion to [M].</span>")
+	to_chat(M, "<span class='notice'>Your mind tingles as you are fed the potion. You can hear radio waves now!</span>")
+	var/obj/item/implant/radio/slime/ncr/imp = new
 	imp.implant(M, user)
 	qdel(src)
